+++ src/options/args.rs
@@ -371,6 +371,14 @@ pub struct ProcessArgs {
         alias = "whole-word"
     )]
     pub whole_word: bool,
+
+    #[arg(
+        long,
+        action = ArgAction::SetTrue,
+        help = "Collapse process tree by default.",
+        alias = "tree-collapse"
+    )]
+    pub tree_collapse: bool,
 }
 
 /// Temperature arguments/config options.