+++ src/options/config/flags.rs
@@ -45,4 +45,5 @@ pub(crate) struct FlagConfig {
     pub(crate) enable_cache_memory: Option<bool>,
     pub(crate) retention: Option<StringOrNum>,
     pub(crate) average_cpu_row: Option<bool>, // FIXME: This makes no sense outside of basic mode, add a basic mode config section.
+    pub(crate) tree_collapse: Option<bool>,
 }